pascals-triangle: update to v1.0.0 tests
